Integrated frontend with backend signup, signin, sendotp
