--- Project Files/frontend/src/pages/SignUp.jsx
@@ -1,83 +1,0 @@
-import React, { useState } from "react";
-
-const SignUp = () => {
-    const [formData, setFormData] = useState({
-        fullName: '',
-        userName: '',
-        email: '',
-        password: ''
-    });
-
-    const handleChange = (e) => {
-        setFormData({
-            ...formData,
-            [e.target.name]: e.target.value
-        });
-    };
-
-    const handleSubmit = async (e) => {
-        e.preventDefault();
-        try {
-            const response = await fetch('http://localhost:5124/user/signup', {
-                method: 'POST',
-                body: JSON.stringify(formData),
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-            });
-
-            if (response.ok) {
-                alert("Sign up successful!");
-            } else {
-                const error = await response.text();
-                alert(`Sign up failed: ${error}`);
-            }
-        } catch (error) {
-            console.error("Sign up error:", error);
-            alert("An error occurred during sign up.");
-        }
-    };
-
-    return (
-        <div>
-            <h2>Sign Up</h2>
-            <form onSubmit={handleSubmit}>
-                <input
-                    type="text"
-                    name="fullName"
-                    value={formData.fullName}
-                    onChange={handleChange}
-                    placeholder="Full Name"
-                    required
-                />
-                <input
-                    type="text"
-                    name="userName"
-                    value={formData.userName}
-                    onChange={handleChange}
-                    placeholder="User Name"
-                    required
-                />
-                <input
-                    type="email"
-                    name="email"
-                    value={formData.email}
-                    onChange={handleChange}
-                    placeholder="Email"
-                    required
-                />
-                <input
-                    type="password"
-                    name="password"
-                    value={formData.password}
-                    onChange={handleChange}
-                    placeholder="Password"
-                    required
-                />
-                <button type="submit">Sign Up</button>
-            </form>
-        </div>
-    );
-};
-
-export default SignUp;
